Enable independent execution of 'main.py'
Reason
In a previous commit, the return type of `main` method of 'main.py' is changed from `void` to `int` to indicate whether the given query propositional sentence / formula holds for the given knowledge base or not.
This change made `main` method unable to execute independently when 'main.py' is executed as a source program ie., `main` method now requires another function or call to complete its desired functionality as per the assignment guidelines.

Solution
This code change will remove that dependency of `main` method to meet the desired requirements of the assignment guidelines.